Add login, logout, cart persistence for visitor to user
Added tests to confirm cart items remain available to vistor after visitor creates
a new account. User sees logout button when logged in and login in button when logged
out. Functionality needs to be implemented throughout the site via the navbar, once
we can merge the frontend and backend code.